# Improved DependenciesReader performance (~10x faster) 