@samuelsainz 2022-05-24
@polac24 If you folks don't mind I'd rather leave some of the performance tests. I think there are still some improvements that we can make in the future—for example regarding the different Set copies and iterations over those Sets. I will take a look at it later and performance tests would be useful for that